+++ tests/live_tests/scenarios/rpki_rtr_example/data6.py
@@ -13,14 +13,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-class BIRD2RPKIRTRScenario_Data4(object):
+class RPKIRTRScenario_Data6(object):
 
     DATA = {
-        "rs_IPAddress":                     "192.0.2.2",
+        "rs_IPAddress":                     "2001:db8:1:1::2",
 
-        "AS10745_allowed_prefixes":         "199.43.0.0/24",
-        "AS3333_allowed_prefixes":          "193.0.0.0/21",
+        "AS10745_allowed_prefixes":         "2001:500:4::/48",
+        "AS3333_allowed_prefixes":          "2001:67c:2e8::/48",
 
-        "AS1_1_IPAddress":                  "192.0.2.111",
-        "AS1_1":                            "193.0.0.0/24"
+        "AS1_1_IPAddress":                  "2001:db8:1:1::111",
+        "AS1_1":                            "2001:67c:2e8::/56"
     }